requirements.txt
nba_api>=1.8.0
pandas>=2.2,<3
numpy>=1.26,<3
scikit-learn>=1.5,<2
streamlit>=1.41,<2
plotly>=5.24,<6
joblib>=1.4,<2
requests>=2.32,<3
python-dotenv>=1.0,<2
pytest>=8.3,<9
